% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skatR.R
\name{SKATOh}
\alias{SKATOh}
\title{Optimal sequence kernel association test (SKAT-O) using marginal variant score statistics}
\usage{
SKATOh(obj, G, W.beta = c(1, 25), rho = c(0, 0.1^2, 0.2^2, 0.3^2, 0.4^2,
  0.5^2, 0.5, 1))
}
\arguments{
\item{obj}{a fitted null model using KAT.null() or KAT.cnull()}

\item{G}{genotype matrix, sample in rows, variant in columns}

\item{W.beta}{Beta parameters for variant weights}

\item{rho}{proportion weight assigned to burden test statistic}
}
\value{
SKAT-O p-value
}
\description{
Efficient SKAT-O p-value calculation using marginal variant score statistics directly
}
\details{
Efficiently compute the SKAT-O significance p-value based on the variant test statistics (typically of much smaller dimension).
The individual p-values of weighted SKAT and burden test are comptued more accurately using the SKATh.
To obtain more accruate results, the one-dimensional integration is computed based on the convolution of survival function of
1-DF chi-square mixtures and 1-DF chi-square density function. For details, please see the Wu et. al (2016) reference.
}
\references{
Lee, S., Wu, M. C., and Lin, X. (2012) Optimal tests for rare variant effects in sequencing association studies. Biostatistics, 13, 762-775.

Wu,B., Guan,W., and Pankow,J.S. (2016) On efficient and accurate calculation of significance p-values for sequence kernel association test of variant set. AHG, 80(2), 123-135.
}
\keyword{SKAT-O}
